Transform solc-select into a python package
