--- experimental/bin/solc
@@ -1,12 +1,0 @@
-#!/usr/bin/env python3
-import os
-import solc_select
-import sys
-
-res = solc_select.current_version()
-if res:
-    (version, _) = res
-    path = f"{solc_select.artifacts_dir}/solc-{version}"
-    os.execv(path, [path] + sys.argv[1:])
-else:
-    sys.exit(1)
