--- scripts/chroot.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-cd $HOST_PWD
-/usr/bin/solc $@
