--- scripts/solc
@@ -1,11 +1,0 @@
-#!/bin/bash
-
-BIN_DIR=$SSELECT_INSTALL_DIR/usr/bin
-
-if [[ "$1" == "--versions" ]]; then
-    $BIN_DIR/solc-select --list
-elif [ -z "$SOLC_VERSION" ]; then
-    $BIN_DIR/solc-default $@
-else
-    $BIN_DIR/solc-v$SOLC_VERSION $@
-fi
